why dont I just ivec3 for each

+++ Chunk.hpp
@@ -62,8 +62,8 @@ public:
 		blocks[x * HEIGHT * DEPTH + y * DEPTH + z] = type;
 	}
 		
-	BlockType GetBlock(glm::vec3 pos) const {
-		return GetBlock(static_cast<int>(pos.x), static_cast<int>(pos.y), static_cast<int>(pos.z));
+	BlockType GetBlock(glm::ivec3 pos) const {
+		return GetBlock(pos.x, pos.y, pos.z);
 	}
 
 	glm::vec2 GetPosition() {
